fix: BinaryHeap elements are dropped twice
